move bit-error from the legacy to Harmony
