+++ .bitmap
@@ -1,6 +1,25 @@
 /* THIS IS A BIT-AUTO-GENERATED FILE. DO NOT EDIT THIS FILE DIRECTLY. */
 
 {
+    "bit-error": {
+        "files": [
+            {
+                "name": "bit-error.ts",
+                "relativePath": "bit-error.ts",
+                "test": false
+            },
+            {
+                "name": "index.ts",
+                "relativePath": "index.ts",
+                "test": false
+            }
+        ],
+        "mainFile": "index.ts",
+        "rootDir": "scopes/harmony/bit-error",
+        "origin": "AUTHORED",
+        "exported": false,
+        "lanes": []
+    },
     "component-id": {
         "files": [
             {
