+++ scopes/component/legacy-bit-id/exceptions/invalid-id-chunk.ts
@@ -1,7 +1,7 @@
 import chalk from 'chalk';
+import { BitError } from '@teambit/bit-error';
 
-// @todo: should extends BitError
-export default class InvalidIdChunk extends Error {
+export default class InvalidIdChunk extends BitError {
   id: string;
 
   constructor(id: string) {
