+++ scopes/defender/tester/exceptions/no-test-files-found.ts
@@ -1,4 +1,4 @@
-import { BitError } from 'bit-bin/dist/error/bit-error';
+import { BitError } from '@teambit/bit-error';
 
 export class NoTestFilesFound extends BitError {
   constructor(testRegex: string) {
