+++ scopes/dependencies/dependency-resolver/dependencies/exceptions/unknown-dep-type.ts
@@ -1,4 +1,4 @@
-import { BitError } from 'bit-bin/dist/error/bit-error';
+import { BitError } from '@teambit/bit-error';
 
 export class UnknownDepType extends BitError {
   constructor(readonly type: string) {
