+++ scopes/dependencies/dependency-resolver/exceptions/main-aspect-not-installable.ts
@@ -1,4 +1,4 @@
-import { BitError } from 'bit-bin/dist/error/bit-error';
+import { BitError } from '@teambit/bit-error';
 
 export class MainAspectNotInstallable extends BitError {
   constructor() {
