+++ scopes/dependencies/dependency-resolver/exceptions/package-manager-not-found.ts
@@ -1,4 +1,4 @@
-import { BitError } from 'bit-bin/dist/error/bit-error';
+import { BitError } from '@teambit/bit-error';
 
 export class PackageManagerNotFound extends BitError {
   constructor(packageManagerName: string) {
