+++ scopes/dependencies/dependency-resolver/exceptions/root-dir-not-defined.ts
@@ -1,4 +1,4 @@
-import { BitError } from 'bit-bin/dist/error/bit-error';
+import { BitError } from '@teambit/bit-error';
 
 export class RootDirNotDefined extends BitError {
   constructor() {
