+++ scopes/explorer/insights/exceptions/insight-already-exists.ts
@@ -1,4 +1,4 @@
-import { BitError } from 'bit-bin/dist/error/bit-error';
+import { BitError } from '@teambit/bit-error';
 
 export default class InsightAlreadyExists extends BitError {
   constructor(readonly insightName: string) {
