+++ scopes/explorer/insights/exceptions/insight-not-found.ts
@@ -1,4 +1,4 @@
-import { BitError } from 'bit-bin/dist/error/bit-error';
+import { BitError } from '@teambit/bit-error';
 
 export default class InsightNotFound extends BitError {
   constructor(readonly insightName: string) {
