+++ scopes/explorer/insights/exceptions/no-data-for-insight.ts
@@ -1,4 +1,4 @@
-import { BitError } from 'bit-bin/dist/error/bit-error';
+import { BitError } from '@teambit/bit-error';
 
 export default class NoDataForInsight extends BitError {
   constructor(readonly insightName: string) {
