+++ scopes/harmony/aspect-loader/exceptions/cannot-load-extension.ts
@@ -1,4 +1,4 @@
-import { BitError } from 'bit-bin/dist/error/bit-error';
+import { BitError } from '@teambit/bit-error';
 
 export class CannotLoadExtension extends BitError {
   constructor(private extensionName: string, private error: Error) {
