+++ scopes/harmony/bit-error/bit-error.ts
