+++ scopes/harmony/cli/exceptions/command-not-found.ts
@@ -1,4 +1,4 @@
-import { BitError } from 'bit-bin/dist/error/bit-error';
+import { BitError } from '@teambit/bit-error';
 import chalk from 'chalk';
 
 export class CommandNotFound extends BitError {
