+++ scopes/harmony/config/exceptions/extension-already-configured.ts
@@ -1,4 +1,4 @@
-import { BitError } from 'bit-bin/dist/error/bit-error';
+import { BitError } from '@teambit/bit-error';
 
 export class ExtensionAlreadyConfigured extends BitError {
   constructor(readonly extensionId: string) {
