+++ scopes/pkg/pkg/exceptions/scope-not-found.ts
@@ -1,4 +1,4 @@
-import { BitError } from 'bit-bin/dist/error/bit-error';
+import { BitError } from '@teambit/bit-error';
 
 export class ScopeNotFound extends BitError {
   constructor(readonly scopePath?: string) {
