+++ scopes/pkg/pkg/publisher.ts
@@ -5,7 +5,7 @@ import { Logger } from '@teambit/logger';
 import { Workspace } from '@teambit/workspace';
 import { BitId, BitIds } from 'bit-bin/dist/bit-id';
 import { ExtensionDataList } from 'bit-bin/dist/consumer/config/extension-data';
-import { BitError } from 'bit-bin/dist/error/bit-error';
+import { BitError } from '@teambit/bit-error';
 import { Scope } from 'bit-bin/dist/scope';
 import mapSeries from 'p-map-series';
 import execa from 'execa';
