+++ scopes/workspace/workspace/component-config-file/exceptions/already-exists.ts
@@ -1,4 +1,4 @@
-import { BitError } from 'bit-bin/dist/error/bit-error';
+import { BitError } from '@teambit/bit-error';
 
 export class AlreadyExistsError extends BitError {
   constructor(filePath: string) {
