+++ scopes/workspace/workspace/exceptions/dependency-type-not-supported-in-policy.ts
@@ -1,4 +1,4 @@
-import { BitError } from 'bit-bin/dist/error/bit-error';
+import { BitError } from '@teambit/bit-error';
 
 export class DependencyTypeNotSupportedInPolicy extends BitError {
   constructor(type: string) {
