+++ scopes/workspace/workspace/exceptions/incorrect-env-aspect.ts
@@ -1,4 +1,4 @@
-import { BitError } from 'bit-bin/dist/error/bit-error';
+import { BitError } from '@teambit/bit-error';
 
 export class IncorrectEnvAspect extends BitError {
   constructor(id: string, envType: string, envId: string) {
